+++ src/main/java/com/atlauncher/gui/dialogs/ImportInstanceDialog.java
@@ -190,7 +190,7 @@ public class ImportInstanceDialog extends JDialog {
                 setVisible(true);
                 DialogManager.okDialog().setTitle(GetText.tr("Failed To Add Pack"))
                         .setContent(new HTMLBuilder().center().text(GetText.tr(
-                                "An error occured when trying to add Curse pack.<br/><br/>Check the console for more information."))
+                                "An error occured when trying to add CurseForge pack.<br/><br/>Check the console for more information."))
                                 .build())
                         .setType(DialogManager.ERROR).show();
             } else {